Refactor isBulkDomainTransfer
