+++ client/my-sites/checkout/checkout-thank-you/redesign-v2/sections/Product.tsx
@@ -1,3 +1,4 @@
+import { isBulkDomainTransfer } from '../../utils';
 import DomainsTransferredList from './product/DomainsTransferredList';
 import ProductPlan, { ProductPlanProps } from './product/ProductPlan';
 import type { ReceiptPurchase } from 'calypso/state/receipts/types';
@@ -7,9 +8,8 @@ const Product = ( {
 	primaryPurchase,
 	siteID,
 	purchases,
-	isBulkDomainTransfer,
-}: ProductPlanProps & { purchases: ReceiptPurchase[]; isBulkDomainTransfer: boolean } ) => {
-	if ( isBulkDomainTransfer ) {
+}: ProductPlanProps & { purchases: ReceiptPurchase[] } ) => {
+	if ( isBulkDomainTransfer( purchases ) ) {
 		return <DomainsTransferredList purchases={ purchases } />;
 	}
