+++ client/my-sites/checkout/checkout-thank-you/utils.ts
@@ -5,6 +5,7 @@ import {
 	JETPACK_SOCIAL_PRODUCTS,
 	JETPACK_SEARCH_PRODUCTS,
 	JETPACK_VIDEOPRESS_PRODUCTS,
+	isDomainTransfer,
 } from '@automattic/calypso-products';
 import JetpackBackupPluginImage from 'calypso/assets/images/jetpack/jetpack-plugin-image-backup.svg';
 import JetpackBoostPluginImage from 'calypso/assets/images/jetpack/jetpack-plugin-image-boost.svg';
@@ -13,6 +14,7 @@ import JetpackSocialPluginImage from 'calypso/assets/images/jetpack/jetpack-plug
 import JetpackVideopressPluginImage from 'calypso/assets/images/jetpack/jetpack-plugin-image-videopress.svg';
 import JetpackPluginImage from 'calypso/assets/images/jetpack/licensing-activation-plugin-install.svg';
 import { domainManagementEdit, domainManagementList } from 'calypso/my-sites/domains/paths';
+import type { ReceiptPurchase } from 'calypso/state/receipts/types';
 
 const buildKeyValuePairByProductSlugs = (
 	productSlugs: ReadonlyArray< string >,
@@ -86,4 +88,8 @@ export function getDomainManagementUrl(
 	domain: string | undefined
 ): string {
 	return domain ? domainManagementEdit( slug, domain ) : domainManagementList( slug );
+}
+
+export function isBulkDomainTransfer( purchases: ReceiptPurchase[] ): boolean {
+	return purchases?.every( isDomainTransfer );
 }
